Cost Assignment and Creation Rules
Move menu now allows the following:
-You can create a move based on the list that generates results
-Cost starts at the preset modifier
-Can change positions for moves to start, but requires certain rules (depending on trajectory chosen)
-etc.

Everything is working. Can be more modular, but this is still just a prototype before going live. Next step is to setup saving logic, and then
create the assign move menu.

Saving needs a UI experience for users to choose save slots and to delete or edit existing moves.